requirements.txt
langchain
langchain-groq
langgraph
streamlit
python-dotenv
ipython